# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  command.CreateLoan:
    properties:
      creditDesc:
        type: string
      creditRate:
        type: number
      creditSum:
        type: number
      creditTerm:
        type: integer
      creditTitle:
        type: string
    required:
    - creditDesc
    - creditRate
    - creditSum
    - creditTerm
    - creditTitle
    type: object
  command.CreateLoanCounteroffer:
    properties:
      creditRate:
        type: number
      creditTerm:
        type: integer
    required:
    - creditRate
    - creditTerm
    type: object
...
  eliftech-school_backend_internal_controllers.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  eliftech-school_backend_internal_controllers.statusResponse:
    properties:
      status:
        type: string
    type: object
  eliftech-school_backend_internal_controllers.stripeURLResponse:
    properties:
      url:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Back end for 50 Cent Application
  title: 50 Cent API
  version: "1.1"
paths:
...
  /loans:
    post:
      consumes:
      - application/json
      description: Create loan request form
      operationId: createLoan
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.CreateLoan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Loan
      tags:
      - Loans
  /loans/:id/accept:
    post:
      consumes:
      - application/json
      description: Accept loan request
      operationId: acceptLoan
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.stripeURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Accept Loan
      tags:
      - Loans
  /loans/:id/repay:
    post:
      consumes:
      - application/json
      description: Make monthly payment for your loan
      operationId: repayLoan
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.stripeURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Repay loan
      tags:
      - Loans
  /loans/{id}:
    delete:
      consumes:
      - application/json
      description: Delete loan by id
      operationId: loandelete
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Loan
      tags:
      - Loans
    get:
      consumes:
      - application/json
      description: Get loan by id (yet not )
      operationId: getloanbyid
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get loans by id
      tags:
      - Loans
    post:
      consumes:
      - application/json
      description: Update loan by id
      operationId: loanupdate
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.CreateLoan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Loan
      tags:
      - Loans
  /loans/{id}/counteroffers:
    get:
      consumes:
      - application/json
      description: Retrieve all counteroffers related to this loan
      operationId: getLoanCounteroffers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      summary: Loan counteroffers
      tags:
      - Loan
      - Investor
    post:
      consumes:
      - application/json
      description: Investor can suggest his loan terms to consumer
      operationId: loanCounteroffer
      parameters:
      - description: counteroffer terms
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.CreateLoanCounteroffer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      summary: Loan counteroffer
      tags:
      - Loan
      - Investor
  /loans/{id}/transactions:
    get:
      consumes:
      - application/json
      description: User can get transactions related to his loan
      operationId: getLoanTransactions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      summary: Get Loan Transactions
      tags:
      - Loan
      - Investor
      - Consumer
  /loans/{page}/{page_size}:
    get:
      consumes:
      - application/json
      description: Get all loans list
      operationId: getallloans
      parameters:
      - description: Current Page
        in: path
        name: page
        required: true
        type: integer
      - description: Page Size
        in: path
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Loans
      tags:
      - Loans
  /loans/counteroffers/{id}/accept:
    post:
      consumes:
      - application/json
      description: Consumer can accept investor's terms and update loan details
      operationId: loanCounterofferAccept
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      summary: Loan counteroffer accept
      tags:
      - Loan
      - Investor
      - Consumer
  /loans/counteroffers/{id}/reject:
    post:
      consumes:
      - application/json
      description: Consumer can reject investor's terms if he doesn't want to accept
        them
      operationId: loanCounterofferReject
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      summary: Loan counteroffer reject
      tags:
      - Loan
      - Investor
      - Consumer
  /stripe/webhook:
    post:
      consumes:
      - application/json
      description: Webhook for stripe events such as account confirmation, payment
        success or fail
      operationId: stripeWebhook
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/eliftech-school_backend_internal_controllers.errorResponse'
      summary: Webhook for stripe events
      tags:
      - Webhook
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"